% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_tpf.R
\name{augment_tpf}
\alias{augment_tpf}
\title{Augment TPF}
\usage{
augment_tpf(data, touchpoints = NULL)
}
\arguments{
\item{data}{data}

\item{touchpoints}{touchpoints}
}
\value{
data
}
\description{
Augment TPF
}
