% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_transition_matrix.R
\name{compose_transition_matrix}
\alias{compose_transition_matrix}
\title{Compose a Random Markov Chain Transition Matrix}
\usage{
compose_transition_matrix(unique_tchp_type)
}
\arguments{
\item{unique_tchp_type}{a character vector of unique possible touch-points in path to conversion}
}
\value{
a transition matrix
}
\description{
This function allows you to generate a random markov chain transition matrix following the form
required for the conversion path analysis i.e. Square matrix with two absorbing states (conv, drop),
a (start) state that can't be reached from any other state, and a user defined set of unique states
representing possible touch points on path to conversion.
}
