% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_flow.R
\name{conversion_flow}
\alias{conversion_flow}
\title{Simulate Paths and Generate Conversion Flow Diagram}
\usage{
conversion_flow(transition_matrix, num_steps, num_sim)
}
\arguments{
\item{transition_matrix}{transition_matrix}

\item{num_steps}{num_steps}

\item{num_sim}{num_sim}
}
\value{
a list
}
\description{
Given a transition matrix run a simulation to generate `num_sim` number of paths with `num_steps`.
This function will return an aggregated table of transition points as well as the conversion flow diagram.
}
