% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_flow_diagram.R
\name{conversion_flow_diagram}
\alias{conversion_flow_diagram}
\title{Simulate Paths and Generate Conversion Flow Diagram}
\usage{
conversion_flow_diagram(transition_matrix, num_steps, num_sim)
}
\arguments{
\item{transition_matrix}{transition_matrix}

\item{num_steps}{num_steps}

\item{num_sim}{num_sim}
}
\value{
a list
}
\description{
Simulate Paths and Generate Conversion Flow Diagram
}
