% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_conversion_path.R
\docType{data}
\name{digital_conversion_path}
\alias{digital_conversion_path}
\title{Simulated Paths to Conversion with Digital Touchpoints}
\format{
A data frame with 676 rows and 3 variables:
\describe{
  \item{path}{charecter string with touchpoints in sequence separted by >}
  \item{conv_count}{count of conversions from a given path}
  \item{drop_count}{count of non-conversions from a given path}
}
}
\usage{
digital_conversion_path
}
\description{
A dataset containing paths to conversion and counts of conversion and non-conversion for each path
}
\keyword{datasets}
