% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_path_list.R
\name{extract_path_list}
\alias{extract_path_list}
\title{Extract List of Paths from Data}
\usage{
extract_path_list(path_data, remove_repeating = FALSE)
}
\arguments{
\item{path_data}{path_data}

\item{remove_repeating}{remove_repeating}
}
\value{
a list of paths
}
\description{
Extract List of Paths from Data
}
