% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_transition_matrix.R
\name{fit_transition_matrix}
\alias{fit_transition_matrix}
\title{Compute MLE Markov Chain Transition Matrix}
\usage{
fit_transition_matrix(path_list, conv_count = NULL, drop_count = NULL)
}
\arguments{
\item{path_list}{path_list}

\item{conv_count}{conv_count}

\item{drop_count}{drop_count}
}
\value{
a transition matrix
}
\description{
Compute MLE Markov Chain Transition Matrix
}
