% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_path_data.R
\name{generate_path_data}
\alias{generate_path_data}
\title{Generate Path Data from Path List}
\usage{
generate_path_data(path_list)
}
\arguments{
\item{path_list}{path_list}
}
\value{
data.frame
}
\description{
Generate Path Data from Path List
}
