% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.proper_transition_matrix.R
\name{is.proper_transition_matrix}
\alias{is.proper_transition_matrix}
\title{Test Format of Transition Matrix}
\usage{
is.proper_transition_matrix(transition_matrix)
}
\arguments{
\item{transition_matrix}{a transiton matrix to be tested}
}
\value{
a transition matrix
}
\description{
Test if a markov chain transition matrix follows the form
required for the conversion path analysis i.e. Square matrix with two absorbing states (conv, drop),
a (start) state that can't be reached from any other state.
}
