% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_length.R
\name{path_length}
\alias{path_length}
\title{Compute Path Length}
\usage{
path_length(path, remove_repeating = FALSE)
}
\arguments{
\item{path}{vector of paths}

\item{remove_repeating}{Should repeated touchpoints be removed first?}
}
\value{
vector of cleaned paths
}
\description{
Remove repeated touch-points in a path if they are adjacent to one another.
}
