% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_touchpoint_frequency.R
\name{plot_touchpoint_frequency}
\alias{plot_touchpoint_frequency}
\title{Plot Touch-point Frequency}
\usage{
plot_touchpoint_frequency(path_data, remove_repeating = FALSE)
}
\arguments{
\item{path_data}{path_data}
}
\value{
plotly plot
}
\description{
Plot Touch-point Frequency
}
