% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_transition_matrix.R
\name{plot_transition_matrix}
\alias{plot_transition_matrix}
\title{Plot Markov Chain Transition Matrix}
\usage{
plot_transition_matrix(transition_matrix, full = FALSE)
}
\arguments{
\item{transition_matrix}{transition_matrix}
}
\value{
plotly plot
}
\description{
Plot Markov Chain Transition Matrix
}
