% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_repeated.R
\name{remove_repeated}
\alias{remove_repeated}
\title{Remove Repeated Touch-points}
\usage{
remove_repeated(path)
}
\arguments{
\item{path}{vector of paths}
}
\value{
vector of cleaned paths
}
\description{
Remove repeated touch-points in a path if they are adjacent to one another.
}
