% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_path.R
\name{simulate_path}
\alias{simulate_path}
\title{Simulate one Path From Transition Matrix}
\usage{
simulate_path(transition_matrix, num_steps = 10)
}
\arguments{
\item{transition_matrix}{transition_matrix}

\item{num_steps}{num_steps}
}
\value{
data.frame
}
\description{
Simulate one Path From Transition Matrix
}
