% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_path_list.R
\name{simulate_path_list}
\alias{simulate_path_list}
\title{Simulate Path List From Transition Matrix}
\usage{
simulate_path_list(transition_matrix, num_sim, lambda = 5)
}
\arguments{
\item{transition_matrix}{transition_matrix}

\item{num_sim}{number of simulations paths to generate}

\item{lambda}{mean of a Poisson(lambda) distribution to sample path length from}
}
\value{
list
}
\description{
From a given transition matrix simulate a list of paths
}
