% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_path_table.R
\name{simulate_path_table}
\alias{simulate_path_table}
\title{Simulate Path Table From Transition Matrix}
\usage{
simulate_path_table(transition_matrix, num_steps, num_sim)
}
\arguments{
\item{transition_matrix}{transition_matrix}

\item{num_steps}{num_steps}

\item{num_sim}{num_sim}
}
\value{
data.frame
}
\description{
Simulate Path Table From Transition Matrix
}
