% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{tf}
\alias{tf}
\title{Touch-point Frequency Across Paths}
\usage{
tf(paths, touchpoint)
}
\arguments{
\item{paths}{paths}

\item{touchpoint}{touchpoint}
}
\value{
numeric vector
}
\description{
Touch-point Frequency Across Paths
}
