% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpf.R
\name{tpf}
\alias{tpf}
\title{Touch-point Frequency per Path}
\usage{
tpf(paths, touchpoint)
}
\arguments{
\item{paths}{paths}

\item{touchpoint}{touchpoint}
}
\value{
numeric vector
}
\description{
Touch-point Frequency per Path
}
