% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix_from_markov_model.R
\name{transition_matrix_from_markov_model}
\alias{transition_matrix_from_markov_model}
\title{Derive a Proper Transition Matrix from MC Model}
\usage{
transition_matrix_from_markov_model(mc_model)
}
\arguments{
\item{path_list}{mc_model}
}
\value{
a transition matrix
}
\description{
Derive a Proper Transition Matrix from MC Model
}
