% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_transition_matrix.R
\name{compose_transition_matrix}
\alias{compose_transition_matrix}
\title{Compose a Random Markov Chain Transition Matrix}
\usage{
compose_transition_matrix(unique_tchp_type)
}
\arguments{
\item{unique_tchp_type}{unique_tchp_type}
}
\value{
a transition matrix
}
\description{
Compose a Random Markov Chain Transition Matrix
}
