% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_transition_matrix.R
\name{fit_transition_matrix}
\alias{fit_transition_matrix}
\title{Compute MLE Markov Chain Transition Matrix}
\usage{
fit_transition_matrix(path_data)
}
\arguments{
\item{path_data}{a list of paths}
}
\value{
a transition matrix
}
\description{
This function estimates a transition matrix using MLE.
Given a list of paths and two numeric vectors with a number of total conversions and total cases that do not leaad to a conversion.
}
