% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_lengths.R
\name{plot_path_lengths}
\alias{plot_path_lengths}
\title{Plot Path Lengths}
\usage{
plot_path_lengths(path_list)
}
\arguments{
\item{path_list}{path_list}
}
\value{
plotly plot
}
\description{
Plot Path Lengths
}
