% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_touchpoint_frequency.R
\name{plot_touchpoint_frequency}
\alias{plot_touchpoint_frequency}
\title{Plot Touchpoint Frequency}
\usage{
plot_touchpoint_frequency(path_list)
}
\arguments{
\item{path_list}{path_list}
}
\value{
plotly plot
}
\description{
Plot Touchpoint Frequency
}
