% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rep_tchp.R
\name{remove_rep_tchp}
\alias{remove_rep_tchp}
\title{Remove Repeated Touchpoints}
\usage{
remove_rep_tchp(long_path)
}
\arguments{
\item{long_path}{long_path}
}
\value{
short path
}
\description{
Remove Repeated Touchpoints
}
